+++ frontend/Container.js
@@ -21,7 +21,7 @@ var TabbedPane = require('./TabbedPane');
 var assign = require('object-assign');
 
 import type MenuItem from './ContextMenu';
-import type {Base16Theme} from './Themes/Base16Theme';
+import type {Base16Theme} from './Themes/Themes';
 
 type Props = {};
