+++ frontend/DataView/DataView.js
@@ -10,7 +10,7 @@
  */
 'use strict';
 
-import type {Base16Theme} from '../Themes/Base16Theme';
+import type {Base16Theme} from '../Themes/Themes';
 import type {DOMEvent} from '../types';
 
 var React = require('react');
