+++ frontend/PropState.js
@@ -21,7 +21,7 @@ var assign = require('object-assign');
 var decorate = require('./decorate');
 var invariant = require('./invariant');
 
-import type {Base16Theme} from './Themes/Base16Theme';
+import type {Base16Theme} from './Themes/Themes';
 
 class PropState extends React.Component {
   context: {
